requirements.txt
absl-py==0.12.0
anyio==3.1.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
bleach==3.3.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.4
cryptography==3.4.7
cycler==0.10.0
cytoolz==0.11.0
dask==2021.5.0
decorator==5.0.9
defusedxml==0.7.1
dlib==19.22.0
entrypoints==0.3
fsspec==0.9.0
google-auth==1.21.3
google-auth-oauthlib==0.4.1
grpcio==1.38.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imbalanced-learn==0.8.0
importlib-metadata==4.0.1
imutils==0.5.4
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.7.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.5.2
kiwisolver==1.3.1
locket==0.2.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.2.1
mkl-service==2.3.0
mlxtend==0.18.0
nbclassic==0.2.8
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5
notebook==6.4.0
numpy==1.20.2
oauthlib==3.1.1
olefile==0.46
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.2
parso==0.8.2
partd==1.2.0
pickleshare==0.7.5
Pillow==8.2.0
pip==21.0.1
playsound==1.2.2
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyreadline==2.1
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.16.2
scikit-learn==0.24.2
scipy==1.6.3
Send2Trash==1.5.0
setuptools==52.0.0.post20210125
six==1.15.0
sniffio==1.2.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
terminado==0.9.4
testpath==0.5.0
threadpoolctl==2.1.0
toolz==0.11.1
torch==1.8.1
torchaudio==0.8.1
torchinfo==0.1.4
torchvision==0.2.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==2.0.1
wheel==0.36.2
win-inet-pton==1.1.0
wincertstore==0.2
zipp==3.4.1
